feat: Rename 에러 방지 위해
- vue/multi-word-component-names 에러
