+++ client/src/App.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { RouterView } from 'vue-router'
 
-import Header from './components/Header.vue'
-import Footer from '@/components/Footer.vue'
-import SideBar from '@/components/SideBar.vue'
+import Header from '@/components/Header-Item.vue'
+import Footer from '@/components/Footer-Item.vue'
+import SideBar from '@/components/Sidebar-Item.vue'
 </script>
 
 <template>
